+++ android/src/main/java/com/dataxad/flutter_mailer/MethodCallHandlerImpl.java
@@ -1,34 +1,45 @@
-package com.dataxad.fluttermailer;
+package com.dataxad.flutter_mailer;
+
 
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
+
 import android.net.Uri;
 import android.os.Build;
-import androidx.core.content.FileProvider;
 import android.text.Html;
 import android.text.Spanned;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.content.FileProvider;
+
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
+import io.flutter.plugin.common.PluginRegistry;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
-import io.flutter.embedding.engine.plugins.activity.ActivityAware;
-import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
-import io.flutter.plugin.common.MethodCall;
-import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import io.flutter.plugin.common.MethodChannel.Result;
-import io.flutter.plugin.common.PluginRegistry;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
+class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler, PluginRegistry.ActivityResultListener {
+    static class FlutterMailerException extends Exception {
+        final String errorMessage;
+        final Object errorDetails;
+        final String errorCode;
 
-/**
- * FlutterMailerPlugin
- */
-public class FlutterMailerPlugin implements MethodCallHandler, PluginRegistry.ActivityResultListener {
+        FlutterMailerException(String errorCode, @Nullable String errorMessage, @Nullable Object errorDetails) {
+            super();
+            this.errorCode = errorCode;
+            this.errorMessage = errorMessage;
+            this.errorDetails = errorDetails;
+        }
+    }
+
+    private static final String TAG = "FLUTTER_MAILER";
     private static final String IS_HTML = "isHTML";
     private static final String SUBJECT = "subject";
     private static final String BODY = "body";
@@ -38,54 +49,74 @@ public class FlutterMailerPlugin implements MethodCallHandler, PluginRegistry.Ac
     private static final String ATTACHMENTS = "attachments";
     private static final String MAILTO_SCHEME = "mailto:";
     private static final String APP_SCHEMA = "appSchema";
-    private final Registrar mRegistrar;
+    private static final int MAIL_ACTIVITY_REQUEST_CODE = 564;
 
-    private Result mResult;
-    private Activity mActivty;
-    private static final int MAIL_ACTIVTY_REQUEST_CODE = 564;
+    private final Context context;
+    private Activity activity;
+    private MethodChannel.Result mResult;
 
-    /**
-     * Plugin registration.
-     */
-    public static void registerWith(Registrar registrar) {
-        final MethodChannel channel = new MethodChannel(registrar.messenger(), "flutter_mailer");
-        FlutterMailerPlugin plugin = new FlutterMailerPlugin(registrar);
-        channel.setMethodCallHandler(plugin);
-        registrar.addActivityResultListener(plugin);
+    MethodCallHandlerImpl(@NonNull Context context, @Nullable Activity activity) {
+        this.context = context;
+        this.activity = activity;
+    }
+
+    void setActivity(Activity activity) {
+        this.activity = activity;
     }
 
     @Override
-    public void onMethodCall(MethodCall call, Result result) {
+    public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result result) {
         if (call.method.equals("send")) {
-            mail(call, result);
+            mResult = result;
+            try {
+                final Intent intent = mail(call);
+                activity.startActivityForResult(intent, MAIL_ACTIVITY_REQUEST_CODE);
+            } catch (FlutterMailerException e) {
+                result.error(e.errorCode, e.errorMessage, e.errorDetails);
+                mResult = null;
+            } catch (Exception e) {
+                Log.e(TAG, e.getMessage());
+                result.error("UNKNOWN", e.getMessage(), null);
+                mResult = null;
+            }
+        } else if (call.method.equals("isAppInstalled")) {
+            if (call.hasArgument(APP_SCHEMA) && call.argument(APP_SCHEMA) != null && isAppInstalled((String) call.argument(APP_SCHEMA))) {
+                result.success(true);
+            } else {
+                result.success(false);
+            }
         } else {
             result.notImplemented();
         }
     }
 
-    private FlutterMailerPlugin(Registrar registrar) {
-        this.mRegistrar = registrar;
-        this.mActivty = registrar.activity();
+    @Override
+    public boolean onActivityResult(int requestCode, int resultCode, Intent intent) {
+        if (requestCode == MAIL_ACTIVITY_REQUEST_CODE && mResult != null) {
+            mResult.success(null);
+            return false;
+        }
+        return false;
     }
 
 
-    private void mail(MethodCall options, Result callback) {
-        Context context = mRegistrar.context();
+    private Intent mail(MethodCall options) throws FlutterMailerException {
+
         Intent intent = new Intent(Intent.ACTION_SENDTO,
                 Uri.parse(MAILTO_SCHEME));
 
 //        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
         if (options.hasArgument(SUBJECT)) {
-            String subject = options.argument(SUBJECT) ;
+            String subject = options.argument(SUBJECT);
             intent.putExtra(Intent.EXTRA_SUBJECT, subject != null ? subject : "");
         }
 
         if (options.hasArgument(BODY)) {
             final String body = options.argument(BODY);
-            CharSequence text =  body != null ? body: "" ;
+            CharSequence text = body != null ? body : "";
 
-            if (options.hasArgument(IS_HTML) && (boolean)options.argument(IS_HTML)) {
+            if (options.hasArgument(IS_HTML) && (boolean) options.argument(IS_HTML)) {
                 text = fromHtml((String) text);
             }
             intent.putExtra(Intent.EXTRA_TEXT, text);
@@ -99,7 +130,7 @@ public class FlutterMailerPlugin implements MethodCallHandler, PluginRegistry.Ac
 
         if (options.hasArgument(CCRecipients)) {
             ArrayList<String> ccRecipients = options.argument(CCRecipients);
-            final String[] r = readableArrayToStringArray( ccRecipients != null ? ccRecipients : new ArrayList<String>() );
+            final String[] r = readableArrayToStringArray(ccRecipients != null ? ccRecipients : new ArrayList<String>());
             intent.putExtra(Intent.EXTRA_CC, r);
         }
 
@@ -112,19 +143,17 @@ public class FlutterMailerPlugin implements MethodCallHandler, PluginRegistry.Ac
         if (options.hasArgument(ATTACHMENTS)) {
             ArrayList<String> attachments = options.argument(ATTACHMENTS);
             if (attachments == null) {
-                callback.error("Attachments_null", "Attachments cannot be null", null);
-            } else if (!attachments.isEmpty() ) {
+                throw new FlutterMailerException("Attachments_null", "Attachments cannot be null", null);
+            } else if (!attachments.isEmpty()) {
                 ArrayList<Uri> uris = new ArrayList<>();
-
 
                 for (int j = 0; j < attachments.size(); j++) {
                     final String path = attachments.get(j);
 
-
                     intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
                     File file = new File(path);
 
-                    final Uri p = FileProvider.getUriForFile(context, mRegistrar.context().getPackageName() + ".adv_provider", file);
+                    final Uri p = FileProvider.getUriForFile(context, context.getPackageName() + ".adv_provider", file);
                     uris.add(p);
                 }
 
@@ -139,44 +168,17 @@ public class FlutterMailerPlugin implements MethodCallHandler, PluginRegistry.Ac
         List<ResolveInfo> list = manager.queryIntentActivities(intent, 0);
 
         if (list == null || list.size() == 0) {
-            Log.e("Flutter_mailer ERROR: ", "size is null or Zero");
-            callback.error("not_available", "no email Managers available", null);
-            return;
+            Log.e(TAG, "size is null or Zero");
+            throw new FlutterMailerException("not_available", "no email Managers available", null);
         }
 
         if (list.size() == 1) {
-//            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            mResult = callback;
-            try {
-                mActivty.startActivityForResult(intent, MAIL_ACTIVTY_REQUEST_CODE);
-            } catch (Exception ex) {
-                Log.e("Flutter_mailer Size==1", ex.getMessage());
-                callback.error("error", ex.getMessage(), null);
-            }
-        } else if (options.hasArgument(APP_SCHEMA) && options.argument(APP_SCHEMA) != null && isAppInstalled((String) options.argument(APP_SCHEMA)) ) {
-            mResult = callback;
-            try {
-                intent.setPackage((String)options.argument(APP_SCHEMA));
-                mActivty.startActivityForResult(intent, MAIL_ACTIVTY_REQUEST_CODE);
-            } catch (Exception ex) {
-                Log.e("Flutter_mailer ERROR: ", ex.getMessage());
-                callback.error("error", ex.getMessage(), null);
-            }
-
-        } else {
-
-            // Intent chooser = Intent.createChooser(intent, "Send Mail");
-            // chooser.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            mResult = callback;
-
-
-            try {
-                mActivty.startActivityForResult(intent, MAIL_ACTIVTY_REQUEST_CODE);
-            } catch (Exception ex) {
-                Log.e("Flutter_mailer ERROR: ", ex.getMessage());
-                callback.error("error", ex.getMessage(), null);
-            }
+            return intent;
+        } else if (options.hasArgument(APP_SCHEMA) && options.argument(APP_SCHEMA) != null && isAppInstalled((String) options.argument(APP_SCHEMA))) {
+            intent.setPackage((String) options.argument(APP_SCHEMA));
+            return intent;
         }
+        return intent;
     }
 
     /**
@@ -200,13 +202,12 @@ public class FlutterMailerPlugin implements MethodCallHandler, PluginRegistry.Ac
     /**
      * Ask the package manager if the app is installed on the device.
      *
-     * @param id    The app id.
-     * @return      true if yes otherwise false.
+     * @param id The app id.
+     * @return true if yes otherwise false.
      */
-    private boolean isAppInstalled (String id) {
-
+    private boolean isAppInstalled(String id) {
         try {
-            mRegistrar.activeContext().getPackageManager().getPackageInfo(id, 0);
+            context.getPackageManager().getPackageInfo(id, 0);
             return true;
         } catch (PackageManager.NameNotFoundException e) {
             return false;
@@ -221,14 +222,4 @@ public class FlutterMailerPlugin implements MethodCallHandler, PluginRegistry.Ac
             return Html.fromHtml(source);
         }
     }
-
-    @Override
-    public boolean onActivityResult(int requestCode, int resultCode, Intent intent) {
-        if(requestCode == MAIL_ACTIVTY_REQUEST_CODE && mResult != null){
-            mResult.success(null);
-            return false;
-        }
-        return false;
-    }
-
 }