+++ lib/flutter_mailer.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:flutter/services.dart';
 
@@ -7,6 +8,32 @@ class FlutterMailer {
 
   static Future<void> send(MailOptions mailOptions) async {
     return _channel.invokeMethod('send', mailOptions.toJson());
+  }
+
+  /// returns true if an app schema is installed on the device
+  /// other wise returns false
+  ///
+  /// ### _Android only_
+  /// returns false on other platforms
+  static Future<bool> isAppInstalled(String schema) async {
+    if (!Platform.isAndroid) {
+      return false;
+    }
+
+    return _channel.invokeMethod('isAppInstalled', <String, String> {'appSchema': schema});
+  }
+
+  /// returns true if can MFMailComposeViewController canSendMail is true
+  /// other wise returns false
+  ///
+  /// ### _IOS only_
+  /// returns false on other platforms
+  static Future<bool> canSendMail() async {
+    if (!Platform.isIOS) {
+      return false;
+    }
+
+    return _channel.invokeMethod('canSendMail');
   }
 }